Tyres_management_project/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
